RSS212-57 - Fix pagination on the Admin Vault page

+++ datavault-common/src/main/java/org/datavaultplatform/common/model/dao/VaultDAO.java
@@ -24,9 +24,9 @@ public interface VaultDAO {
 
     public int getRetentionPolicyCount(int status);
 
-	public Long getTotalNumberOfVaults(String userId);
+	public int getTotalNumberOfVaults(String userId);
 
-	public Long getTotalNumberOfVaults(String userId, String query);
+	public int getTotalNumberOfVaults(String userId, String query);
 	
 	public List<Object[]> getAllProjectsSize();
 }
